requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
typing_extensions==4.7.1
django-environ==0.10.0
Pillow==10.0.0
psycopg2==2.9.6
django-sitetree==1.17.3
django-allauth==0.54.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-currentuser==0.6.1
pytz==2023.3.post1
gunicorn==20.1.0